EnumZipper improvements.
