+++ build.sbt
@@ -179,7 +179,7 @@ lazy val util = crossProject(JVMPlatform, JSPlatform)
   .crossType(CrossType.Pure)
   .in(file("modules/util"))
   .settings(commonSettings)
-  .dependsOn(model)
+  .dependsOn(model, testkit)
   .settings(
     name := "gpp-core-util",
     libraryDependencies ++= Seq(
